move picker to builder container
