+++ src/Pages/ADT3DScenePage/ADT3DScenePage.scss
@@ -100,8 +100,8 @@
 
 .cb-scene-page-mode-toggle-container {
     position: absolute;
-    right: 20px;
-    top: 20px;
+    right: 8px;
+    top: 0px;
     z-index: 1;
 }
