+++ src/Pages/ADT3DScenePage/Internal/ADT3DSceneBuilderContainer.tsx
@@ -1,9 +1,15 @@
-import React from 'react';
+import React, { useReducer } from 'react';
 import BaseComponent from '../../../Components/BaseComponent/BaseComponent';
 import { ADT3DScenePageModes } from '../../../Models/Constants/Enums';
 import ADT3DViewer from '../../../Components/ADT3DViewer/ADT3DViewer';
 import ADT3DSceneBuilder from '../../../Components/ADT3DSceneBuilder/ADT3DSceneBuilder';
 import { IADT3DSceneBuilderProps } from '../ADT3DScenePage.types';
+import FloatingScenePageModeToggle from './FloatingScenePageModeToggle';
+import { SET_ADT_SCENE_PAGE_MODE } from '../../../Models/Constants/ActionTypes';
+import {
+    ADT3DScenePageReducer,
+    defaultADT3DScenePageState
+} from '../ADT3DScenePage.state';
 
 export const ADT3DSceneBuilderContainer: React.FC<IADT3DSceneBuilderProps> = ({
     mode = ADT3DScenePageModes.BuildScene,
@@ -15,6 +21,19 @@ export const ADT3DSceneBuilderContainer: React.FC<IADT3DSceneBuilderProps> = ({
     localeStrings,
     refetchConfig
 }) => {
+    const [state, dispatch] = useReducer(
+        ADT3DScenePageReducer,
+        defaultADT3DScenePageState
+    );
+
+    const handleScenePageModeChange = (
+        newScenePageMode: ADT3DScenePageModes
+    ) => {
+        dispatch({
+            type: SET_ADT_SCENE_PAGE_MODE,
+            payload: newScenePageMode
+        });
+    };
     return (
         <BaseComponent
             theme={theme}
@@ -41,6 +60,11 @@ export const ADT3DSceneBuilderContainer: React.FC<IADT3DSceneBuilderProps> = ({
                     />
                 </div>
             )}
+            <FloatingScenePageModeToggle
+                scene={scene}
+                handleScenePageModeChange={handleScenePageModeChange}
+                selectedMode={state.scenePageMode}
+            />
         </BaseComponent>
     );
 };
